+++ core-ng-test/src/test/java/core/framework/test/redis/MockRedisTest.java
@@ -1,8 +1,11 @@
 package core.framework.test.redis;
 
+import core.framework.api.util.Lists;
 import org.junit.Assert;
 import org.junit.Before;
 import org.junit.Test;
+
+import java.util.List;
 
 /**
  * @author neo
@@ -21,5 +24,14 @@ public class MockRedisTest {
 
         String value = redis.get("key");
         Assert.assertEquals("value", value);
+    }
+
+    @Test
+    public void mget() {
+        redis.set("key1", "value1");
+        redis.set("key2", "value2");
+
+        List<String> values = redis.mget(Lists.newArrayList("key1", "key3", "key2"));
+        Assert.assertEquals(Lists.newArrayList("value1", null, "value2"), values);
     }
 }